rough draft of exercise completed. Missing styles and 'further study' components

+++ app.py
@@ -1,7 +1,7 @@
 from flask import Flask, request, redirect, render_template, url_for, flash, jsonify, session
 from flask_debugtoolbar import DebugToolbarExtension
 from sqlalchemy import desc
-from models import db, User, connect_db
+from models import db, User, Feedback, connect_db
 from flask_bcrypt import Bcrypt
 from forms import AddUserForm, LoginForm, FeedbackForm
 
@@ -14,13 +14,19 @@ debug = DebugToolbarExtension(app)
 bcrypt = Bcrypt()
 connect_db(app)
 
+#TODO clean up view functions by moving most of the logic to some sort of controller
+
 @app.route('/')
 def root():
     """Redirect to /register"""
+    if 'username' in session:
+        return redirect(url_for('show_logged_in', username=session['username']))
     return redirect(url_for('register'))
 
 @app.route('/register', methods=['GET', 'POST'])
 def register():
+    if 'username' in session:
+        return redirect(url_for('show_logged_in', username=session['username']))
     form = AddUserForm()
     if form.validate_on_submit():
         username = form.username.data
@@ -51,15 +57,15 @@ def show_login():
 
 @app.route('/users/<username>')
 def show_logged_in(username):
-    # need to add logic for showing Feedback from user
     if 'username' not in session: 
         flash("Please log in first to view this")
         return redirect('/login')
-    user = User.query.get(username)
-    print(user.username)
+    user = User.query.get(session['username'])
     if user == None:
         return redirect('/login')
-    return render_template('logged_in.html',user=user)
+    feedback = Feedback.query.filter_by(username=session['username']).all()
+    form = FeedbackForm()
+    return render_template('logged_in.html',user=user, form=form, feedback=feedback)
 
 @app.route('/logout', methods=['POST'])
 def process_logout():
@@ -71,26 +77,67 @@ def redirect_logout():
     """Only accept logout requests made via post"""
     return redirect(url_for('show_login'))
 
-# routes to work on tomorrow
-# need to add FeedbackForm
-
 @app.route('/users/<username>/delete', methods=['POST'])
 def delete_account(username):
     return redirect('/')
 
-@app.route('users/<username>/feedback/add', methods=['GET', 'POST'])
+@app.route('/users/<username>/feedback/add', methods=['POST'])
 def handle_feedback_form(username):
-    form = FeedbackForm()
-    return render_template('add_feedback.html')
+    if 'username' not in session:
+        flash("Please log in to post feedback")
+        return redirect(url_for('show_login'))
+    if username != session['username']:
+        flash("You can't post feedback for somebody else's account!")
+        return redirect(url_for('show_logged_in'))
 
-@app.route('feedback/<feedback_id>/update', methods=['GET', 'POST'])
-def handle_feedback_update(feedback_id):
     form = FeedbackForm()
+    if form.validate_on_submit():
+        title = form.title.data
+        content = form.content.data
+    new_feedback = Feedback(title=title,content=content,username=username)
+    db.session.add(new_feedback)
+    db.session.commit()
+    flash("Your feedback has been added")
     return redirect(url_for('show_logged_in', username=session['username']))
 
-@app.route('feedback/<feedback_id>/delete', methods=['POST'])
+@app.route('/feedback/<feedback_id>/update', methods=['GET', 'POST'])
+def handle_feedback_update(feedback_id):
+    if 'username' not in session:
+        flash("Please log in to post feedback")
+        return redirect(url_for('show_login'))
+
+    feedback = Feedback.query.get(feedback_id)
+    if feedback == None:
+        return redirect(url_for('show_logged_in', username=session['username']))
+    if feedback.username != session['username']:
+        flash("You can't update feedback for somebody else's account!")
+        return redirect(url_for('show_logged_in', username=session['username']))
+
+
+    form = FeedbackForm()
+    if form.validate_on_submit():
+        feedback.title = form.title.data
+        feedback.content = feedback.content.data
+        flash("Your feedback was updated")
+        return redirect(url_for('show_logged_in', username=session['username']))
+    return render_template('edit_feedback.html', form=form)
+
+@app.route('/feedback/<feedback_id>/delete', methods=['POST'])
 def delete_feedback(feedback_id):
-    feedback = db.session.get(feedback_id)
+    if 'username' not in session:
+        flash("Please log in to delete feedback")
+        return redirect(url_for('show_login'))
+    feedback = Feedback.query.get(feedback_id)
+    if feedback == None:
+        return redirect(url_for('show_logged_in', username=session['username']))
+    if feedback.username != session['username']:
+        flash("You can't delete feedback for somebody else's account!")
+        return redirect(url_for('show_logged_in', username=session['username']))
     db.session.delete(feedback)
     db.session.commit()
+    flash("Your post has been deleted")
     return redirect(url_for('show_logged_in', username=session['username']))
+
+# def check_if_logged_in():
+#     if 'username' in session:
+#         redirect(url_for('show_logged_in', username=session['username']))
